src/requirements.txt
bandit
Flask
flake8
google-cloud-firestore
google-cloud-logging
pylama
pylint
pytest
pytest-cov
